completed remaining functions

+++ internal/models/subscription.go
@@ -7,7 +7,7 @@ import (
 type Subscription struct {
 	ID          uint      `json:"id" gorm:"primaryKey;autoIncrement"`
 	ServiceName string    `json:"service_name" gorm:"type:varchar(255);not null;index"`
-	Price       int       `json:"price" gorm:"not null"`
+	Price       int64     `json:"price" gorm:"not null"`
 	UserID      string    `json:"user_id" gorm:"type:uuid;not null;index"`
 	StartDate   string    `json:"start_date" gorm:"type:varchar(7);not null;index"`
 	EndDate     *string   `json:"end_date,omitempty" gorm:"type:varchar(7);default:null"`
